fixed verifier unit tests
