+++ data-plane/core/auth/src/testutils.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use std::sync::Once;
+use std::time::{SystemTime, UNIX_EPOCH};
 
 use aws_lc_rs::encoding::AsDer;
 use aws_lc_rs::signature::KeyPair; // Import the KeyPair trait for public_key() method
@@ -17,10 +18,11 @@ static RUSTLS: Once = Once::new();
 
 pub fn initialize_crypto_provider() {
     RUSTLS.call_once(|| {
-        // Set aws-lc as default crypto provider
-        rustls::crypto::aws_lc_rs::default_provider()
-            .install_default()
-            .unwrap();
+        // Set aws-lc as default crypto provider, but don't panic if it's already set
+        if let Err(_) = rustls::crypto::aws_lc_rs::default_provider().install_default() {
+            // Provider might already be installed, which is fine
+            // Just ignore the error and continue
+        }
     });
 }
 
@@ -197,7 +199,11 @@ pub async fn setup_test_jwt_resolver(algorithm: Algorithm) -> (String, MockServe
             "jwks_uri": jwks_uri,
             "response_types_supported": ["code"],
             "subject_types_supported": ["public"],
-            "id_token_signing_alg_values_supported": ["RS256"]
+            "id_token_signing_alg_values_supported": ["RS256"],
+            "scopes_supported": ["openid", "profile", "email"],
+            "token_endpoint_auth_methods_supported": ["client_secret_basic", "client_secret_post"],
+            "claims_supported": ["sub", "iss", "aud", "exp", "iat"],
+            "grant_types_supported": ["authorization_code", "client_credentials"]
         })))
         .mount(&mock_server)
         .await;
@@ -233,4 +239,88 @@ fn bytes_to_pem(key_bytes: &[u8], header: &str, footer: &str) -> String {
     }
 
     format!("{}{}{}", header, pem_body, footer)
+}
+
+/// Test claims structure for JWT testing
+#[derive(Debug, serde::Serialize, serde::Deserialize, PartialEq)]
+pub struct TestClaims {
+    pub sub: String,
+    pub iss: String,
+    pub aud: String,
+    pub exp: u64,
+    pub iat: u64,
+}
+
+impl TestClaims {
+    /// Create new test claims with proper timestamps
+    pub fn new(sub: impl Into<String>, iss: impl Into<String>, aud: impl Into<String>) -> Self {
+        let now = SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .unwrap()
+            .as_secs();
+
+        Self {
+            sub: sub.into(),
+            iss: iss.into(),
+            aud: aud.into(),
+            exp: now + 3600, // Expires in 1 hour
+            iat: now,
+        }
+    }
+}
+
+/// Setup a mock OIDC server for testing both token provider and verifier
+pub async fn setup_oidc_mock_server() -> (MockServer, String, String) {
+    let mock_server = MockServer::start().await;
+    let issuer_url = mock_server.uri();
+
+    // Mock OIDC discovery endpoint
+    Mock::given(method("GET"))
+        .and(path("/.well-known/openid-configuration"))
+        .respond_with(ResponseTemplate::new(200).set_body_json(json!({
+            "issuer": issuer_url,
+            "authorization_endpoint": format!("{}/auth", issuer_url),
+            "token_endpoint": format!("{}/oauth2/token", issuer_url),
+            "jwks_uri": format!("{}/oauth2/jwks.json", issuer_url),
+            "userinfo_endpoint": format!("{}/userinfo", issuer_url),
+            "response_types_supported": ["code"],
+            "subject_types_supported": ["public"],
+            "id_token_signing_alg_values_supported": ["RS256"],
+            "scopes_supported": ["openid", "profile", "email"],
+            "token_endpoint_auth_methods_supported": ["client_secret_basic", "client_secret_post"],
+            "claims_supported": ["sub", "iss", "aud", "exp", "iat"],
+            "grant_types_supported": ["authorization_code", "client_credentials"]
+        })))
+        .mount(&mock_server)
+        .await;
+
+    // Mock token endpoint for client credentials
+    Mock::given(method("POST"))
+        .and(path("/oauth2/token"))
+        .respond_with(ResponseTemplate::new(200).set_body_json(json!({
+            "access_token": "test-access-token-12345",
+            "token_type": "Bearer",
+            "expires_in": 3600
+        })))
+        .mount(&mock_server)
+        .await;
+
+    // Mock JWKS endpoint (needed for token verification)
+    Mock::given(method("GET"))
+        .and(path("/oauth2/jwks.json"))
+        .respond_with(ResponseTemplate::new(200).set_body_json(json!({
+            "keys": [{
+                "kty": "RSA",
+                "kid": "test-key-1",
+                "n": "0vx7agoebGcQSuuPiLJXZptN9nndrQmbXEps2aiAFbWhM78LhWx4cbbfAAtVT86zwu1RK7aPFFxuhDR1L6tSoc_BJECPebWKRXjBZCiFV4n3oknjhMstn64tZ_2W-5JsGY4Hc5n9yBXArwl93lqt7_RN5w6Cf0h4QyQ5v-65YGjQR0_FDW2QvzqY368QQMicAtaSqzs8KJZgnYb9c7d0zgdAZHzu6qMQvRL5hajrn1n91CbOpbIS_4q3eFD1LTXfhHQjEZzXpk2zb7fF_xxGLmNr8zXczK8TGLLcgPYEEYnNJhJRs5vJ3dNb02f1_Q-sNHd8qXe5s7eXs2E4RbQJvQ",
+                "e": "AQAB",
+                "alg": "RS256"
+            }]
+        })))
+        .mount(&mock_server)
+        .await;
+
+    let access_token = "test-access-token-12345".to_string();
+
+    (mock_server, issuer_url, access_token)
 }
